Make a single source of truth for target strings
