+++ crates/cli/src/lib.rs
@@ -20,6 +20,7 @@ use roc_load::{ExpectMetadata, Threading};
 use roc_mono::ir::OptLevel;
 use roc_packaging::cache::RocCacheDir;
 use roc_packaging::tarball::Compression;
+use roc_target::Target;
 use std::env;
 use std::ffi::{CString, OsStr, OsString};
 use std::io;
@@ -28,11 +29,8 @@ use std::os::raw::{c_char, c_int};
 use std::path::{Path, PathBuf};
 use std::process;
 use std::time::Instant;
-use strum::{EnumIter, IntoEnumIterator, IntoStaticStr};
-use target_lexicon::{Aarch64Architecture, BinaryFormat};
-use target_lexicon::{
-    Architecture, Environment, OperatingSystem, Triple, Vendor, X86_32Architecture,
-};
+use strum::IntoEnumIterator;
+use target_lexicon::{Architecture, Triple};
 #[cfg(not(target_os = "linux"))]
 use tempfile::TempDir;
 
@@ -1274,118 +1272,4 @@ fn run_wasm<I: Iterator<Item = S>, S: AsRef<[u8]>>(wasm_path: &std::path::Path, 
 #[cfg(not(feature = "run-wasm32"))]
 fn run_wasm<I: Iterator<Item = S>, S: AsRef<[u8]>>(_wasm_path: &std::path::Path, _args: I) {
     println!("Running wasm files is not supported on this target.");
-}
-
-#[derive(Debug, Copy, Clone, EnumIter, IntoStaticStr, PartialEq, Eq, Default)]
-pub enum Target {
-    #[strum(serialize = "system")]
-    #[default]
-    System,
-    #[strum(serialize = "linux-x86-32")]
-    LinuxX32,
-    #[strum(serialize = "linux-x86-64")]
-    LinuxX64,
-    #[strum(serialize = "linux-arm-64")]
-    LinuxArm64,
-    #[strum(serialize = "macos-x86-64")]
-    MacX64,
-    #[strum(serialize = "macos-arm-64")]
-    MacArm64,
-    #[strum(serialize = "windows-x86-64")]
-    WinX64,
-    #[strum(serialize = "wasm-32")]
-    Wasm32,
-}
-
-const MACOS: OperatingSystem = OperatingSystem::MacOSX {
-    major: 12,
-    minor: 0,
-    patch: 0,
-};
-
-impl Target {
-    pub fn to_triple(self) -> Triple {
-        match self {
-            Target::System => Triple::host(),
-            Target::LinuxX32 => Triple {
-                architecture: Architecture::X86_32(X86_32Architecture::I386),
-                vendor: Vendor::Unknown,
-                operating_system: OperatingSystem::Linux,
-                environment: Environment::Unknown,
-                binary_format: BinaryFormat::Elf,
-            },
-            Target::LinuxX64 => Triple {
-                architecture: Architecture::X86_64,
-                vendor: Vendor::Unknown,
-                operating_system: OperatingSystem::Linux,
-                environment: Environment::Unknown,
-                binary_format: BinaryFormat::Elf,
-            },
-            Target::LinuxArm64 => Triple {
-                architecture: Architecture::Aarch64(Aarch64Architecture::Aarch64),
-                vendor: Vendor::Unknown,
-                operating_system: OperatingSystem::Linux,
-                environment: Environment::Unknown,
-                binary_format: BinaryFormat::Elf,
-            },
-            Target::WinX64 => Triple {
-                architecture: Architecture::X86_64,
-                vendor: Vendor::Unknown,
-                operating_system: OperatingSystem::Windows,
-                environment: Environment::Gnu,
-                binary_format: BinaryFormat::Coff,
-            },
-            Target::MacX64 => Triple {
-                architecture: Architecture::X86_64,
-                vendor: Vendor::Apple,
-                operating_system: MACOS,
-                environment: Environment::Unknown,
-                binary_format: BinaryFormat::Macho,
-            },
-            Target::MacArm64 => Triple {
-                architecture: Architecture::Aarch64(Aarch64Architecture::Aarch64),
-                vendor: Vendor::Apple,
-                operating_system: MACOS,
-                environment: Environment::Unknown,
-                binary_format: BinaryFormat::Macho,
-            },
-            Target::Wasm32 => Triple {
-                architecture: Architecture::Wasm32,
-                vendor: Vendor::Unknown,
-                operating_system: OperatingSystem::Wasi,
-                environment: Environment::Unknown,
-                binary_format: BinaryFormat::Wasm,
-            },
-        }
-    }
-}
-
-impl From<&Target> for Triple {
-    fn from(target: &Target) -> Self {
-        target.to_triple()
-    }
-}
-
-impl std::fmt::Display for Target {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}", Into::<&'static str>::into(self))
-    }
-}
-
-impl std::str::FromStr for Target {
-    type Err = String;
-
-    fn from_str(string: &str) -> Result<Self, Self::Err> {
-        match string {
-            "system" => Ok(Target::System),
-            "linux-x86-32" => Ok(Target::LinuxX32),
-            "linux-x86-64" => Ok(Target::LinuxX64),
-            "linux-arm-64" => Ok(Target::LinuxArm64),
-            "macos-x86-64" => Ok(Target::MacX64),
-            "macos-arm-64" => Ok(Target::MacArm64),
-            "windows-x86-64" => Ok(Target::WinX64),
-            "wasm-32" => Ok(Target::Wasm32),
-            _ => Err(format!("Roc does not know how to compile to {}", string)),
-        }
-    }
 }
